Foreign.CUDA.BLAS.Sparse.Analysis

Description

Documentation

newtype Info

An opaque structure holding the information collected in the analysis phase of the solution of the sparse triangular linear system.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsesolveanalysisinfot

Constructors

Info

Fields

useInfo :: Ptr ()

createInfo :: IO Info

Create and initialise the solve and analysis structure to default values.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsecreatesolveanalysisinfo

destroyInfo :: Info -> IO ()

Release memory associated with a matrix solver structure.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsedestroysolveanalysisinfo

newtype Info_csrsv2

http://docs.nvidia.com/cuda/cusparse/index.html#csrsv2infot

Constructors

Info_csrsv2

Fields

useInfo_csrsv2 :: Ptr ()

createInfo_csrsv2 :: IO Info_csrsv2

destroyInfo_csrsv2 :: Info_csrsv2 -> IO ()

newtype Info_csric02

http://docs.nvidia.com/cuda/cusparse/index.html#csric02infot

Constructors

Info_csric02

Fields

useInfo_csric02 :: Ptr ()

createInfo_csric02 :: IO Info_csric02

destroyInfo_csric02 :: Info_csric02 -> IO ()

newtype Info_csrilu02

http://docs.nvidia.com/cuda/cusparse/index.html#csrilu02infot

Constructors

Info_csrilu02

Fields

useInfo_csrilu02 :: Ptr ()

createInfo_csrilu02 :: IO Info_csrilu02

destroyInfo_csrilu02 :: Info_csrilu02 -> IO ()

newtype Info_bsrsv2

http://docs.nvidia.com/cuda/cusparse/index.html#bsrsv2infot

Constructors

Info_bsrsv2

Fields

useInfo_bsrsv2 :: Ptr ()

createInfo_bsrsv2 :: IO Info_bsrsv2

destroyInfo_bsrsv2 :: Info_bsrsv2 -> IO ()

newtype Info_bsrsm2

http://docs.nvidia.com/cuda/cusparse/index.html#bsrsvminfot

Constructors

Info_bsrsm2

Fields

useInfo_bsrsm2 :: Ptr ()

createInfo_bsrsm2 :: IO Info_bsrsm2

destroyInfo_bsrsm2 :: Info_bsrsm2 -> IO ()

newtype Info_bsric02

http://docs.nvidia.com/cuda/cusparse/index.html#bsric02infot

Constructors

Info_bsric02

Fields

useInfo_bsric02 :: Ptr ()

createInfo_bsric02 :: IO Info_bsric02

destroyInfo_bsric02 :: Info_bsric02 -> IO ()

newtype Info_bsrilu02

http://docs.nvidia.com/cuda/cusparse/index.html#bsrilu02infot

Constructors

Info_bsrilu02

Fields

useInfo_bsrilu02 :: Ptr ()

createInfo_bsrilu02 :: IO Info_bsrilu02

destroyInfo_bsrilu02 :: Info_bsrilu02 -> IO ()

newtype Info_csrgemm2

http://docs.nvidia.com/cuda/cusparse/index.html#csrgemm2infot

Constructors

Info_csrgemm2

Fields

useInfo_csrgemm2 :: Ptr ()

createInfo_csrgemm2 :: IO Info_csrgemm2

destroyInfo_csrgemm2 :: Info_csrgemm2 -> IO ()

newtype Info_csrsm2

https://docs.nvidia.com/cuda/cusparse/index.html#csrsm2infot

Constructors

Info_csrsm2

Fields

useInfo_csrsm2 :: Ptr ()

createInfo_csrsm2 :: IO Info_csrsm2

destroyInfo_csrsm2 :: Info_csrsm2 -> IO ()

newtype Info_color

Constructors

Info_color

Fields

useInfo_color :: Ptr ()

createInfo_color :: IO Info_color

destroyInfo_color :: Info_color -> IO ()

newtype Info_csru2csr

Constructors

Info_csru2csr

Fields

useInfo_csru2csr :: Ptr ()

createInfo_csru2csr :: IO Info_csru2csr

destroyInfo_csru2csr :: Info_csru2csr -> IO ()

newtype Info_prune

Constructors

Info_prune

Fields

useInfo_prune :: Ptr ()

createInfo_prune :: IO Info_prune

destroyInfo_prune :: Info_prune -> IO ()