Foreign.CUDA.BLAS.Sparse.Context

Description

Context management

newtype Handle

An opaque handle to the cuSPARSE library context, which is passed to all library function calls.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsehandlet

Constructors

Handle

Fields

useHandle :: Ptr ()

create :: IO Handle

This function initializes the cuSPARSE library and creates a handle on the cuSPARSE context. It must be called before any other cuSPARSE API function is invoked. It allocates hardware resources necessary for accessing the GPU.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsecreate

destroy :: Handle -> IO ()

This function releases CPU-side resources used by the cuSPARSE library. The release of GPU-side resources may be deferred until the application shuts down.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsedestroy

Utilities

data PointerMode

For functions which take scalar value arguments, determines whether those values are passed by reference on the host or device.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsepointermode_t

Constructors

Host

Device

setPointerMode :: Handle -> PointerMode -> IO ()

Set the pointer mode used by cuSPARSE library functions.

The default mode is for values to be passed by reference from the host.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsesetpointermode

getPointerMode :: Handle -> IO PointerMode

Get the pointer mode used by cuSPARSE library functions to pass scalar arguments.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsegetpointermode