Foreign.CUDA.BLAS.Sparse.Convert

Description

For more information see the cuSPARSE function reference:

http://docs.nvidia.com/cuda/cusparse/index.html#cusparse-format-conversion-reference

Documentation

data Direction

Indicates the underlying storage model for elements of matrices.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsedirectiont

Constructors

Row

Column

data Action

This type indicates whether the operation is performed only on indices (Symbolic) or on data and indices (Numeric).

http://docs.nvidia.com/cuda/cusparse/index.html#cusparseactiont

Constructors

Symbolic

Numeric

data Hybrid

An opaque structure holding the matrix in hybrid (HYB) format.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsehybmatt

data HybridPartition

Indicates how to perform the partitioning of the matrix into regular (ELL) and irregular (COO) parts of the HYB format.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsehybpartitiont

Constructors

Auto

User

Max

data Info_csru2csr

sbsr2csr :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

dbsr2csr :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

cbsr2csr :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

zbsr2csr :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

sgebsr2gebsc_bufferSize :: Handle -> Int -> Int -> Int -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> IO Int

dgebsr2gebsc_bufferSize :: Handle -> Int -> Int -> Int -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> IO Int

cgebsr2gebsc_bufferSize :: Handle -> Int -> Int -> Int -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> IO Int

zgebsr2gebsc_bufferSize :: Handle -> Int -> Int -> Int -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> IO Int

sgebsr2gebsc :: Handle -> Int -> Int -> Int -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Action -> IndexBase -> DevicePtr () -> IO ()

dgebsr2gebsc :: Handle -> Int -> Int -> Int -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Action -> IndexBase -> DevicePtr () -> IO ()

cgebsr2gebsc :: Handle -> Int -> Int -> Int -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Action -> IndexBase -> DevicePtr () -> IO ()

zgebsr2gebsc :: Handle -> Int -> Int -> Int -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Action -> IndexBase -> DevicePtr () -> IO ()

sgebsr2gebsr_bufferSize :: Handle -> Direction -> Int -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> Int -> Int -> IO Int

dgebsr2gebsr_bufferSize :: Handle -> Direction -> Int -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> Int -> Int -> IO Int

cgebsr2gebsr_bufferSize :: Handle -> Direction -> Int -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> Int -> Int -> IO Int

zgebsr2gebsr_bufferSize :: Handle -> Direction -> Int -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> Int -> Int -> IO Int

sgebsr2gebsr :: Handle -> Direction -> Int -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> DevicePtr () -> IO ()

dgebsr2gebsr :: Handle -> Direction -> Int -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> DevicePtr () -> IO ()

cgebsr2gebsr :: Handle -> Direction -> Int -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> DevicePtr () -> IO ()

zgebsr2gebsr :: Handle -> Direction -> Int -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> DevicePtr () -> IO ()

xgebsr2gebsrNnz :: Handle -> Direction -> Int -> Int -> Int -> MatrixDescriptor -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> MatrixDescriptor -> DevicePtr Int32 -> Int -> Int -> Ptr Int32 -> DevicePtr () -> IO ()

sgebsr2csr :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

dgebsr2csr :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

cgebsr2csr :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

zgebsr2csr :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

scsr2gebsr_bufferSize :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> IO Int

dcsr2gebsr_bufferSize :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> IO Int

ccsr2gebsr_bufferSize :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> IO Int

zcsr2gebsr_bufferSize :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> IO Int

scsr2gebsr :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> DevicePtr () -> IO ()

dcsr2gebsr :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> DevicePtr () -> IO ()

ccsr2gebsr :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> DevicePtr () -> IO ()

zcsr2gebsr :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Int -> DevicePtr () -> IO ()

xcsr2gebsrNnz :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr Int32 -> DevicePtr Int32 -> MatrixDescriptor -> DevicePtr Int32 -> Int -> Int -> Ptr Int32 -> DevicePtr () -> IO ()

xcoo2csr :: Handle -> DevicePtr Int32 -> Int -> Int -> DevicePtr Int32 -> IndexBase -> IO ()

scsc2dense :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Float -> Int -> IO ()

dcsc2dense :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Double -> Int -> IO ()

ccsc2dense :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr (Complex Float) -> Int -> IO ()

zcsc2dense :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr (Complex Double) -> Int -> IO ()

scsc2hyb :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Hybrid -> Int -> HybridPartition -> IO ()

dcsc2hyb :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Hybrid -> Int -> HybridPartition -> IO ()

ccsc2hyb :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Hybrid -> Int -> HybridPartition -> IO ()

zcsc2hyb :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Hybrid -> Int -> HybridPartition -> IO ()

scsr2bsr :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

dcsr2bsr :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

ccsr2bsr :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

zcsr2bsr :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

xcsr2bsrNnz :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> MatrixDescriptor -> DevicePtr Int32 -> Ptr Int32 -> IO ()

xcsr2coo :: Handle -> DevicePtr Int32 -> Int -> Int -> DevicePtr Int32 -> IndexBase -> IO ()

scsr2csc :: Handle -> Int -> Int -> Int -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Action -> IndexBase -> IO ()

dcsr2csc :: Handle -> Int -> Int -> Int -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Action -> IndexBase -> IO ()

ccsr2csc :: Handle -> Int -> Int -> Int -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Action -> IndexBase -> IO ()

zcsr2csc :: Handle -> Int -> Int -> Int -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Action -> IndexBase -> IO ()

scsr2dense :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Float -> Int -> IO ()

dcsr2dense :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Double -> Int -> IO ()

ccsr2dense :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr (Complex Float) -> Int -> IO ()

zcsr2dense :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr (Complex Double) -> Int -> IO ()

scsr2hyb :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Hybrid -> Int -> HybridPartition -> IO ()

dcsr2hyb :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Hybrid -> Int -> HybridPartition -> IO ()

ccsr2hyb :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Hybrid -> Int -> HybridPartition -> IO ()

zcsr2hyb :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Hybrid -> Int -> HybridPartition -> IO ()

sdense2csc :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> Int -> DevicePtr Int32 -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

ddense2csc :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> Int -> DevicePtr Int32 -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

cdense2csc :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> Int -> DevicePtr Int32 -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

zdense2csc :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> Int -> DevicePtr Int32 -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

sdense2csr :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> Int -> DevicePtr Int32 -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

ddense2csr :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> Int -> DevicePtr Int32 -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

cdense2csr :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> Int -> DevicePtr Int32 -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

zdense2csr :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> Int -> DevicePtr Int32 -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

sdense2hyb :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> Int -> DevicePtr Int32 -> Hybrid -> Int -> HybridPartition -> IO ()

ddense2hyb :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> Int -> DevicePtr Int32 -> Hybrid -> Int -> HybridPartition -> IO ()

cdense2hyb :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> Int -> DevicePtr Int32 -> Hybrid -> Int -> HybridPartition -> IO ()

zdense2hyb :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> Int -> DevicePtr Int32 -> Hybrid -> Int -> HybridPartition -> IO ()

shyb2csc :: Handle -> MatrixDescriptor -> Hybrid -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

dhyb2csc :: Handle -> MatrixDescriptor -> Hybrid -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

chyb2csc :: Handle -> MatrixDescriptor -> Hybrid -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

zhyb2csc :: Handle -> MatrixDescriptor -> Hybrid -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

shyb2csr :: Handle -> MatrixDescriptor -> Hybrid -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

dhyb2csr :: Handle -> MatrixDescriptor -> Hybrid -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

chyb2csr :: Handle -> MatrixDescriptor -> Hybrid -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

zhyb2csr :: Handle -> MatrixDescriptor -> Hybrid -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> IO ()

shyb2dense :: Handle -> MatrixDescriptor -> Hybrid -> DevicePtr Float -> Int -> IO ()

dhyb2dense :: Handle -> MatrixDescriptor -> Hybrid -> DevicePtr Double -> Int -> IO ()

chyb2dense :: Handle -> MatrixDescriptor -> Hybrid -> DevicePtr (Complex Float) -> Int -> IO ()

zhyb2dense :: Handle -> MatrixDescriptor -> Hybrid -> DevicePtr (Complex Double) -> Int -> IO ()

snnz :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> Int -> DevicePtr Int32 -> Ptr Int32 -> IO ()

dnnz :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> Int -> DevicePtr Int32 -> Ptr Int32 -> IO ()

cnnz :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> Int -> DevicePtr Int32 -> Ptr Int32 -> IO ()

znnz :: Handle -> Direction -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> Int -> DevicePtr Int32 -> Ptr Int32 -> IO ()

createIdentityPermutation :: Handle -> Int -> DevicePtr Int32 -> IO ()

xcoosort_bufferSizeExt :: Handle -> Int -> Int -> Int -> DevicePtr Int32 -> DevicePtr Int32 -> IO Int

xcoosortByRow :: Handle -> Int -> Int -> Int -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr () -> IO ()

xcoosortByColumn :: Handle -> Int -> Int -> Int -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr () -> IO ()

xcsrsort_bufferSizeExt :: Handle -> Int -> Int -> Int -> DevicePtr Int32 -> DevicePtr Int32 -> IO Int

xcsrsort :: Handle -> Int -> Int -> Int -> MatrixDescriptor -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr () -> IO ()

xcscsort_bufferSizeExt :: Handle -> Int -> Int -> Int -> DevicePtr Int32 -> DevicePtr Int32 -> IO Int

xcscsort :: Handle -> Int -> Int -> Int -> MatrixDescriptor -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr () -> IO ()

scsru2csr_bufferSizeExt :: Handle -> Int -> Int -> Int -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csru2csr -> IO Int

dcsru2csr_bufferSizeExt :: Handle -> Int -> Int -> Int -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csru2csr -> IO Int

ccsru2csr_bufferSizeExt :: Handle -> Int -> Int -> Int -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csru2csr -> IO Int

zcsru2csr_bufferSizeExt :: Handle -> Int -> Int -> Int -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csru2csr -> IO Int

scsru2csr :: Handle -> Int -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csru2csr -> DevicePtr () -> IO ()

dcsru2csr :: Handle -> Int -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csru2csr -> DevicePtr () -> IO ()

ccsru2csr :: Handle -> Int -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csru2csr -> DevicePtr () -> IO ()

zcsru2csr :: Handle -> Int -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csru2csr -> DevicePtr () -> IO ()

scsr2csru :: Handle -> Int -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csru2csr -> DevicePtr () -> IO ()

dcsr2csru :: Handle -> Int -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csru2csr -> DevicePtr () -> IO ()

ccsr2csru :: Handle -> Int -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csru2csr -> DevicePtr () -> IO ()

zcsr2csru :: Handle -> Int -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csru2csr -> DevicePtr () -> IO ()

csr2cscEx :: Handle -> Int -> Int -> Int -> DevicePtr () -> Type -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr () -> Type -> DevicePtr Int32 -> DevicePtr Int32 -> Action -> IndexBase -> Type -> IO ()

scsr2csr_compress :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> DevicePtr Int32 -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Float -> IO ()

dcsr2csr_compress :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> DevicePtr Int32 -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Double -> IO ()

ccsr2csr_compress :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> DevicePtr Int32 -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Complex Float -> IO ()

zcsr2csr_compress :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> DevicePtr Int32 -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Complex Double -> IO ()