Foreign.CUDA.BLAS.Sparse.Error

Description

Documentation

data Status

Error codes used by cuSPARSE library functions

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsestatust

Constructors

Success

NotInitialized

AllocFailed

InvalidValue

ArchMismatch

MappingError

ExecutionFailed

InternalError

MatrixTypeNotSupported

ZeroPivot

describe :: Status -> String

data CUSparseException

Constructors

ExitCode Status

UserError String

cusparseError :: String -> IO a

Raise a CUSparseException in the IO Monad

resultIfOk :: (Status, a) -> IO a

Return the results of a function on successful execution, otherwise throw an exception with an error string associated with the return code

nothingIfOk :: Status -> IO ()

Throw an exception with an error string associated with an unsuccessful return code, otherwise return unit.

checkStatus :: CInt -> IO ()

Throw an error if given error code is not CUSPARSE_STATUS_SUCCESS