Foreign.CUDA.BLAS.Sparse.Level2

Description

For more information see the cuSPARSE Level-2 function reference:

http://docs.nvidia.com/cuda/cusparse/index.html#cusparse-level-2-function-reference

Documentation

data Operation

Indicates which operations need to be performed with the sparse matrix.

N: no transpose selected

T: transpose operation

C: conjugate transpose

http://docs.nvidia.com/cuda/cusparse/index.html#cusparseoperationt

Constructors

N

T

C

data Direction

Indicates the underlying storage model for elements of matrices.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsedirectiont

Constructors

Row

Column

data Type

Used to specify the type of data underlying a void* pointer. For example, it is used in the routine csrmvEx.

Constructors

R32f

R64f

R16f

R8i

C32f

C64f

C16f

C8i

R8u

C8u

R32i

C32i

R32u

C32u

data Algorithm

Specify the algorithm to use, for example used in the routine csrmvEx.

Constructors

Alg0

Alg1

data Policy

Indicates whether level information is used by some solver algorithms.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsesolvepolicy_t

Constructors

NoLevel

UseLevel

data MatrixDescriptor

An opaque type used to describe the shape and properties of a matrix.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsematdescrt

data Hybrid

An opaque structure holding the matrix in hybrid (HYB) format.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsehybmatt

data Info

An opaque structure holding the information collected in the analysis phase of the solution of the sparse triangular linear system.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsesolveanalysisinfot

data Info_bsrsv2

http://docs.nvidia.com/cuda/cusparse/index.html#bsrsv2infot

data Info_csrsv2

http://docs.nvidia.com/cuda/cusparse/index.html#csrsv2infot

sbsrmv :: Handle -> Direction -> Operation -> Int -> Int -> Int -> Ptr Float -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> DevicePtr Float -> Ptr Float -> DevicePtr Float -> IO ()

dbsrmv :: Handle -> Direction -> Operation -> Int -> Int -> Int -> Ptr Double -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> DevicePtr Double -> Ptr Double -> DevicePtr Double -> IO ()

cbsrmv :: Handle -> Direction -> Operation -> Int -> Int -> Int -> Ptr (Complex Float) -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> DevicePtr (Complex Float) -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> IO ()

zbsrmv :: Handle -> Direction -> Operation -> Int -> Int -> Int -> Ptr (Complex Double) -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> DevicePtr (Complex Double) -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> IO ()

sbsrxmv :: Handle -> Direction -> Operation -> Int -> Int -> Int -> Int -> Ptr Float -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> DevicePtr Float -> Ptr Float -> DevicePtr Float -> IO ()

dbsrxmv :: Handle -> Direction -> Operation -> Int -> Int -> Int -> Int -> Ptr Double -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> DevicePtr Double -> Ptr Double -> DevicePtr Double -> IO ()

cbsrxmv :: Handle -> Direction -> Operation -> Int -> Int -> Int -> Int -> Ptr (Complex Float) -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> DevicePtr (Complex Float) -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> IO ()

zbsrxmv :: Handle -> Direction -> Operation -> Int -> Int -> Int -> Int -> Ptr (Complex Double) -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> DevicePtr (Complex Double) -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> IO ()

scsrmv :: Handle -> Operation -> Int -> Int -> Int -> Ptr Float -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Float -> Ptr Float -> DevicePtr Float -> IO ()

dcsrmv :: Handle -> Operation -> Int -> Int -> Int -> Ptr Double -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Double -> Ptr Double -> DevicePtr Double -> IO ()

ccsrmv :: Handle -> Operation -> Int -> Int -> Int -> Ptr (Complex Float) -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr (Complex Float) -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> IO ()

zcsrmv :: Handle -> Operation -> Int -> Int -> Int -> Ptr (Complex Double) -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr (Complex Double) -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> IO ()

sbsrsv2_bufferSize :: Handle -> Direction -> Operation -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Info_bsrsv2 -> IO Int

dbsrsv2_bufferSize :: Handle -> Direction -> Operation -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Info_bsrsv2 -> IO Int

cbsrsv2_bufferSize :: Handle -> Direction -> Operation -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Info_bsrsv2 -> IO Int

zbsrsv2_bufferSize :: Handle -> Direction -> Operation -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Info_bsrsv2 -> IO Int

sbsrsv2_analysis :: Handle -> Direction -> Operation -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Info_bsrsv2 -> Policy -> DevicePtr () -> IO ()

dbsrsv2_analysis :: Handle -> Direction -> Operation -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Info_bsrsv2 -> Policy -> DevicePtr () -> IO ()

cbsrsv2_analysis :: Handle -> Direction -> Operation -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Info_bsrsv2 -> Policy -> DevicePtr () -> IO ()

zbsrsv2_analysis :: Handle -> Direction -> Operation -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Info_bsrsv2 -> Policy -> DevicePtr () -> IO ()

sbsrsv2_solve :: Handle -> Direction -> Operation -> Int -> Int -> Ptr Float -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Info_bsrsv2 -> DevicePtr Float -> DevicePtr Float -> Policy -> DevicePtr () -> IO ()

dbsrsv2_solve :: Handle -> Direction -> Operation -> Int -> Int -> Ptr Double -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Info_bsrsv2 -> DevicePtr Double -> DevicePtr Double -> Policy -> DevicePtr () -> IO ()

cbsrsv2_solve :: Handle -> Direction -> Operation -> Int -> Int -> Ptr (Complex Float) -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Info_bsrsv2 -> DevicePtr (Complex Float) -> DevicePtr (Complex Float) -> Policy -> DevicePtr () -> IO ()

zbsrsv2_solve :: Handle -> Direction -> Operation -> Int -> Int -> Ptr (Complex Double) -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Int -> Info_bsrsv2 -> DevicePtr (Complex Double) -> DevicePtr (Complex Double) -> Policy -> DevicePtr () -> IO ()

xbsrsv2_zeroPivot :: Handle -> Info_bsrsv2 -> Ptr Int32 -> IO ()

scsrsv_analysis :: Handle -> Operation -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Info -> IO ()

dcsrsv_analysis :: Handle -> Operation -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Info -> IO ()

ccsrsv_analysis :: Handle -> Operation -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Info -> IO ()

zcsrsv_analysis :: Handle -> Operation -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Info -> IO ()

scsrsv_solve :: Handle -> Operation -> Int -> Ptr Float -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Info -> DevicePtr Float -> DevicePtr Float -> IO ()

dcsrsv_solve :: Handle -> Operation -> Int -> Ptr Double -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Info -> DevicePtr Double -> DevicePtr Double -> IO ()

ccsrsv_solve :: Handle -> Operation -> Int -> Ptr (Complex Float) -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Info -> DevicePtr (Complex Float) -> DevicePtr (Complex Float) -> IO ()

zcsrsv_solve :: Handle -> Operation -> Int -> Ptr (Complex Double) -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Info -> DevicePtr (Complex Double) -> DevicePtr (Complex Double) -> IO ()

scsrsv2_bufferSize :: Handle -> Operation -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csrsv2 -> IO Int

dcsrsv2_bufferSize :: Handle -> Operation -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csrsv2 -> IO Int

ccsrsv2_bufferSize :: Handle -> Operation -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csrsv2 -> IO Int

zcsrsv2_bufferSize :: Handle -> Operation -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csrsv2 -> IO Int

scsrsv2_analysis :: Handle -> Operation -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csrsv2 -> Policy -> DevicePtr () -> IO ()

dcsrsv2_analysis :: Handle -> Operation -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csrsv2 -> Policy -> DevicePtr () -> IO ()

ccsrsv2_analysis :: Handle -> Operation -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csrsv2 -> Policy -> DevicePtr () -> IO ()

zcsrsv2_analysis :: Handle -> Operation -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csrsv2 -> Policy -> DevicePtr () -> IO ()

scsrsv2_solve :: Handle -> Operation -> Int -> Int -> Ptr Float -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csrsv2 -> DevicePtr Float -> DevicePtr Float -> Policy -> DevicePtr () -> IO ()

dcsrsv2_solve :: Handle -> Operation -> Int -> Int -> Ptr Double -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csrsv2 -> DevicePtr Double -> DevicePtr Double -> Policy -> DevicePtr () -> IO ()

ccsrsv2_solve :: Handle -> Operation -> Int -> Int -> Ptr (Complex Float) -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csrsv2 -> DevicePtr (Complex Float) -> DevicePtr (Complex Float) -> Policy -> DevicePtr () -> IO ()

zcsrsv2_solve :: Handle -> Operation -> Int -> Int -> Ptr (Complex Double) -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Info_csrsv2 -> DevicePtr (Complex Double) -> DevicePtr (Complex Double) -> Policy -> DevicePtr () -> IO ()

xcsrsv2_zeroPivot :: Handle -> Info_csrsv2 -> Ptr Int32 -> IO ()

shybmv :: Handle -> Operation -> Ptr Float -> MatrixDescriptor -> Hybrid -> DevicePtr Float -> Ptr Float -> DevicePtr Float -> IO ()

dhybmv :: Handle -> Operation -> Ptr Double -> MatrixDescriptor -> Hybrid -> DevicePtr Double -> Ptr Double -> DevicePtr Double -> IO ()

chybmv :: Handle -> Operation -> Ptr (Complex Float) -> MatrixDescriptor -> Hybrid -> DevicePtr (Complex Float) -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> IO ()

zhybmv :: Handle -> Operation -> Ptr (Complex Double) -> MatrixDescriptor -> Hybrid -> DevicePtr (Complex Double) -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> IO ()

shybsv_analysis :: Handle -> Operation -> MatrixDescriptor -> Hybrid -> Info -> IO ()

dhybsv_analysis :: Handle -> Operation -> MatrixDescriptor -> Hybrid -> Info -> IO ()

chybsv_analysis :: Handle -> Operation -> MatrixDescriptor -> Hybrid -> Info -> IO ()

zhybsv_analysis :: Handle -> Operation -> MatrixDescriptor -> Hybrid -> Info -> IO ()

shybsv_solve :: Handle -> Operation -> Ptr Float -> MatrixDescriptor -> Hybrid -> Info -> DevicePtr Float -> DevicePtr Float -> IO ()

dhybsv_solve :: Handle -> Operation -> Ptr Double -> MatrixDescriptor -> Hybrid -> Info -> DevicePtr Double -> DevicePtr Double -> IO ()

chybsv_solve :: Handle -> Operation -> Ptr (Complex Float) -> MatrixDescriptor -> Hybrid -> Info -> DevicePtr (Complex Float) -> DevicePtr (Complex Float) -> IO ()

zhybsv_solve :: Handle -> Operation -> Ptr (Complex Double) -> MatrixDescriptor -> Hybrid -> Info -> DevicePtr (Complex Double) -> DevicePtr (Complex Double) -> IO ()

sgemvi :: Handle -> Operation -> Int -> Int -> Ptr Float -> DevicePtr Float -> Int -> Int -> DevicePtr Float -> DevicePtr Int32 -> Ptr Float -> DevicePtr Float -> IndexBase -> DevicePtr () -> IO ()

dgemvi :: Handle -> Operation -> Int -> Int -> Ptr Double -> DevicePtr Double -> Int -> Int -> DevicePtr Double -> DevicePtr Int32 -> Ptr Double -> DevicePtr Double -> IndexBase -> DevicePtr () -> IO ()

cgemvi :: Handle -> Operation -> Int -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> Int -> DevicePtr (Complex Float) -> DevicePtr Int32 -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> IndexBase -> DevicePtr () -> IO ()

zgemvi :: Handle -> Operation -> Int -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> Int -> DevicePtr (Complex Double) -> DevicePtr Int32 -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> IndexBase -> DevicePtr () -> IO ()

sgemvi_bufferSize :: Handle -> Operation -> Int -> Int -> Int -> IO Int

dgemvi_bufferSize :: Handle -> Operation -> Int -> Int -> Int -> IO Int

cgemvi_bufferSize :: Handle -> Operation -> Int -> Int -> Int -> IO Int

zgemvi_bufferSize :: Handle -> Operation -> Int -> Int -> Int -> IO Int

csrmvEx :: Handle -> Algorithm -> Operation -> Int -> Int -> Int -> Ptr () -> Type -> MatrixDescriptor -> DevicePtr () -> Type -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr () -> Type -> Ptr () -> Type -> DevicePtr () -> Type -> Type -> DevicePtr () -> IO ()

csrmvEx_bufferSize :: Handle -> Algorithm -> Operation -> Int -> Int -> Int -> Ptr () -> Type -> MatrixDescriptor -> DevicePtr () -> Type -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr () -> Type -> Ptr () -> Type -> DevicePtr () -> Type -> Type -> IO Int

scsrmv_mp :: Handle -> Operation -> Int -> Int -> Int -> Ptr Float -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Float -> Ptr Float -> DevicePtr Float -> IO ()

dcsrmv_mp :: Handle -> Operation -> Int -> Int -> Int -> Ptr Double -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Double -> Ptr Double -> DevicePtr Double -> IO ()

ccsrmv_mp :: Handle -> Operation -> Int -> Int -> Int -> Ptr (Complex Float) -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr (Complex Float) -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> IO ()

zcsrmv_mp :: Handle -> Operation -> Int -> Int -> Int -> Ptr (Complex Double) -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr (Complex Double) -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> IO ()

csrsv_analysisEx :: Handle -> Operation -> Int -> Int -> MatrixDescriptor -> DevicePtr () -> Type -> DevicePtr Int32 -> DevicePtr Int32 -> Info -> Type -> IO ()

csrsv_solveEx :: Handle -> Operation -> Int -> Ptr () -> Type -> MatrixDescriptor -> DevicePtr () -> Type -> DevicePtr Int32 -> DevicePtr Int32 -> Info -> DevicePtr () -> Type -> DevicePtr () -> Type -> Type -> IO ()