Foreign.CUDA.BLAS.Sparse.Matrix.Hybrid

Description

Documentation

newtype Hybrid

An opaque structure holding the matrix in hybrid (HYB) format.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsehybmatt

Constructors

Hybrid

Fields

useHYB :: Ptr ()

data HybridPartition

Indicates how to perform the partitioning of the matrix into regular (ELL) and irregular (COO) parts of the HYB format.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsehybpartitiont

Constructors

Auto

User

Max

createHYB :: IO Hybrid

Create a new (opaque) hybrid matrix.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsecreatehybmat

destroyHYB :: Hybrid -> IO ()

Destroy and release any memory associated with a hybrid matrix.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsedestroyhybmat