Foreign.CUDA.BLAS.Sparse.Matrix.Descriptor

Description

Documentation

newtype MatrixDescriptor

An opaque type used to describe the shape and properties of a matrix.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsematdescrt

Constructors

MatrixDescriptor

Fields

useMatDescr :: Ptr ()

data MatrixType

Indicates the type of matrix.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsematrixtypet

Constructors

General

Symmetric

Hermitian

Triangular

data IndexBase

Indicates whether indexing of matrix elements starts at zero or one.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparseindexbaset

Constructors

Zero

One

data Diagonal

Indicates whether the diagonal elements of the matrix are unity. The diagonal elements are always assumed to be present, but if Unit is passed to an API routine, then the routine assumes that all diagonal entries are unity and will not read or modify those entries.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsediagtypet

Constructors

NonUnit

Unit

data Fill

Indicates whether the upper or lower part of the sparse matrix is stored.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsefillmodet

Constructors

Lower

Upper

createMatDescr :: IO MatrixDescriptor

destroyMatDescr :: MatrixDescriptor -> IO ()

Release memory associated with a matrix descriptor.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsedestroymatdescr

getDiagonal :: MatrixDescriptor -> IO Diagonal

Get the Diagonal type field of the matrix descriptor.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsegetmatdiagtype

getFillMode :: MatrixDescriptor -> IO Fill

Get the Fill mode of the matrix descriptor.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsegetmatfillmode

getIndexBase :: MatrixDescriptor -> IO IndexBase

Get the IndexBase mode of the matrix descriptor.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsegetmatindexbase

getMatrixType :: MatrixDescriptor -> IO MatrixType

Get the MatrixType mode of the matrix descriptor.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsegetmattype

setDiagonal :: MatrixDescriptor -> Diagonal -> IO ()

Set the Diagonal type field of the matrix descriptor.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsesetmatdiagtype

setFillMode :: MatrixDescriptor -> Fill -> IO ()

Set the Fill mode of the matrix descriptor.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsesetmatfillmode

setIndexBase :: MatrixDescriptor -> IndexBase -> IO ()

Set the IndexBase mode of the matrix descriptor.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsesetmatindexbase

setMatrixType :: MatrixDescriptor -> MatrixType -> IO ()

Set the MatrixType mode of the matrix descriptor.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsesetmattype